requirements.txt
absl-py
pytorch_gan_metrics
scipy
tensorflow
tensorboardX
tqdm
pandas

--find-links https://download.pytorch.org/whl/cu117/torch_stable.html
torch==1.13.0+cu117  --extra-index-url https://download.pytorch.org/whl/cu117
torchvision==0.14.0+cu117  --extra-index-url https://download.pytorch.org/whl/cu117
torchdata
